requirements.txt
backports.zoneinfo==0.2.1
certifi==2020.12.5
chardet==4.0.0
idna==2.10
mysql-connector-python==8.0.23
numpy==1.20.1
protobuf==3.15.6
pyTelegramBotAPI==3.7.6
requests==2.25.1
six==1.15.0
timezonefinder==5.2.0
urllib3==1.26.4
